+++ Hardware/ISensor.cs
@@ -33,15 +33,15 @@ namespace OpenHardwareMonitor.Hardware {
   }
 
   public struct SensorValue {
-    private readonly float value;
+    private readonly double value;
     private readonly DateTime time;
 
-    public SensorValue(float value, DateTime time) {
+    public SensorValue(double value, DateTime time) {
       this.value = value;
       this.time = time;
     }
 
-    public float Value { get { return value; } }
+    public double Value { get { return value; } }
     public DateTime Time { get { return time; } }
   }
 
@@ -59,9 +59,9 @@ namespace OpenHardwareMonitor.Hardware {
 
     IReadOnlyArray<IParameter> Parameters { get; }
 
-    float? Value { get; }
-    float? Min { get; }
-    float? Max { get; }
+    double? Value { get; }
+    double? Min { get; }
+    double? Max { get; }
 
     void ResetMin();
     void ResetMax();
